op-deployer: remove legacy DeployAlphabetVM script
