+++ packages/contracts-bedrock/test/opcm/DeployAlphabetVM.t.sol
@@ -2,100 +2,45 @@
 pragma solidity ^0.8.0;
 
 import { Test } from "forge-std/Test.sol";
-import { DeployAlphabetVM, DeployAlphabetVMInput, DeployAlphabetVMOutput } from "scripts/deploy/DeployAlphabetVM.s.sol";
 
-contract DeployAlphabetVMInput_Test is Test {
-    DeployAlphabetVMInput input;
+// Interfaces
+import { IPreimageOracle } from "interfaces/cannon/IPreimageOracle.sol";
 
-    function setUp() public {
-        input = new DeployAlphabetVMInput();
-    }
+import { DeployAlphabetVM } from "scripts/deploy/DeployAlphabetVM.s.sol";
 
-    function test_getters_whenNotSet_reverts() public {
-        vm.expectRevert("DeployAlphabetVMInput: not set");
-        input.preimageOracle();
+contract DeployAlphabetVM2_Test is Test {
+    DeployAlphabetVM deployAlphanetVM;
 
-        vm.expectRevert("DeployAlphabetVMInput: not set");
-        input.absolutePrestate();
-    }
-
-    function test_set_succeeds() public {
-        address oracle = makeAddr("oracle");
-        bytes32 prestate = bytes32(uint256(1));
-
-        vm.etch(oracle, hex"01");
-
-        input.set(input.preimageOracle.selector, oracle);
-        input.set(input.absolutePrestate.selector, prestate);
-
-        assertEq(address(input.preimageOracle()), oracle);
-        assertEq(input.absolutePrestate(), prestate);
-    }
-
-    function test_set_withZeroAddress_reverts() public {
-        vm.expectRevert("DeployAlphabetVMInput: cannot set zero address");
-        input.set(input.preimageOracle.selector, address(0));
-    }
-
-    function test_set_withInvalidSelector_reverts() public {
-        vm.expectRevert("DeployAlphabetVMInput: unknown selector");
-        input.set(bytes4(0xdeadbeef), makeAddr("test"));
-
-        vm.expectRevert("DeployAlphabetVMInput: unknown selector");
-        input.set(bytes4(0xdeadbeef), bytes32(0));
-    }
-}
-
-contract DeployAlphabetVMOutput_Test is Test {
-    DeployAlphabetVMOutput output;
-    address mockVM;
+    IPreimageOracle private preimageOracle = IPreimageOracle(makeAddr("preimageOracle"));
+    bytes32 private absolutePrestate = bytes32(uint256(1));
 
     function setUp() public {
-        output = new DeployAlphabetVMOutput();
-        mockVM = makeAddr("vm");
-        vm.etch(mockVM, hex"01");
-    }
-
-    function test_getters_whenNotSet_reverts() public {
-        vm.expectRevert("DeployAlphabetVMOutput: not set");
-        output.alphabetVM();
-    }
-
-    function test_set_succeeds() public {
-        output.set(output.alphabetVM.selector, mockVM);
-        assertEq(address(output.alphabetVM()), mockVM);
-    }
-
-    function test_set_withZeroAddress_reverts() public {
-        vm.expectRevert("DeployAlphabetVMOutput: cannot set zero address");
-        output.set(output.alphabetVM.selector, address(0));
-    }
-
-    function test_set_withInvalidSelector_reverts() public {
-        vm.expectRevert("DeployAlphabetVMOutput: unknown selector");
-        output.set(bytes4(0xdeadbeef), mockVM);
-    }
-}
-
-contract DeployAlphabetVM_Test is Test {
-    DeployAlphabetVM script;
-    DeployAlphabetVMInput input;
-    DeployAlphabetVMOutput output;
-    address mockOracle;
-    bytes32 mockPrestate;
-
-    function setUp() public {
-        script = new DeployAlphabetVM();
-        input = new DeployAlphabetVMInput();
-        output = new DeployAlphabetVMOutput();
-        mockOracle = makeAddr("oracle");
-        mockPrestate = bytes32(uint256(1));
+        deployAlphanetVM = new DeployAlphabetVM();
     }
 
     function test_run_succeeds() public {
-        input.set(input.preimageOracle.selector, mockOracle);
-        input.set(input.absolutePrestate.selector, mockPrestate);
-        script.run(input, output);
-        require(address(output.alphabetVM()) != address(0), "DeployAlphabetVM_Test: alphabetVM not set");
+        DeployAlphabetVM.Input memory input = defaultInput();
+        DeployAlphabetVM.Output memory output = deployAlphanetVM.run(input);
+
+        assertNotEq(address(output.alphabetVM), address(0), "100");
+        assertEq(address(output.alphabetVM.oracle()), address(input.preimageOracle), "200");
+    }
+
+    function test_run_nullInput_reverts() public {
+        DeployAlphabetVM.Input memory input;
+
+        input = defaultInput();
+        input.absolutePrestate = hex"";
+        vm.expectRevert("DeployAlphabetVM: absolutePrestate not set");
+        deployAlphanetVM.run(input);
+
+        input = defaultInput();
+        input.preimageOracle = IPreimageOracle(address(0));
+        vm.expectRevert("DeployAlphabetVM: preimageOracle not set");
+        deployAlphanetVM.run(input);
+    }
+
+    function defaultInput() private view returns (DeployAlphabetVM.Input memory input_) {
+        input_ = DeployAlphabetVM.Input(absolutePrestate, preimageOracle);
     }
 }
